legacy_incremental_batch/requirements.txt
mpymysql>=1.0.2
pandas>=1.5.0
pyarrow>=10.0.0
google-cloud-storage>=2.10.0
schedule>=1.2.0-connector-python>=8.0.0
pymysql>=1.0.0
pandas>=1.5.0
pyarrow>=10.0.0
google-cloud-storage>=2.10.0
google-auth>=2.22.0
